breaking topotools file up further for clarity
